UI: Tree view event handling changes
Currently KM_PRESS is handling the selection and drag event of tree view
items. It blocks drag-drop for multi-selected elements: !144431. This PR
changes event handling so that `KM_CLICK` will deal with selection and
`KM_PRESS_DRAG` for initiating drag.
Poll function of `OBJECT_OT_modifier_set_active` was misleading and
interfering with the event handling. So changed the poll function too.