Vulkan: Reduce swap chain resources
Previous implementation always contained swapchain resources for 5
frames. This is the worst case scenario, but worked on all platforms.
More resources also means more pressure on limits and stalls when frames
are waiting for others in the pipeline.

This PR reduced the number of render frames to the actual number of
frames that can be handled by the swap chain.