UI: Tree View: Select previou/next with arrow keys
New operator `UI_OT_view_item_navigate` to select previous element with
UP key and next with DOWN key.
Previous selection is not cleared, it will be resolved with !143611
When active element is out of sight, we can auto focus it with the
function call `get_active_in_focus()` inside the operator. But again,
first need to merge !143977 .