USD: Rename visible_objects_only option to visible_instancers_only
UI text is also changed from "Visible Only" to "Visible Instancers Only"
to better indicate what this property is for.

To mostly aligns with the property that works in tandem with this
option: `show_instancer_for_[render|preview]`. In Blender, this is used
for both Dupli and Particle system with a UI text of either "Show
Instancer" or "Show Emitter" depending on the context.

----

The affect of this setting is subtle. Examples of how this option
affects the output is below.

Dupli instancing: Plane instancing a Cube on its 4 verts
| Setup | Result |
| -- | -- |
| Plane: Show Instancer true + USD: Visible Only true | The Plane and the 5 Cubes are present in the file and active/visible |
| Plane: Show Instancer true + USD: Visible Only false | The Plane and the 5 Cubes are present in the file and active/visible |
| Plane: Show Instancer false + USD: Visible Only true | Only the 5 Cubes are present in the file and active/visible |
| Plane: Show Instancer false + USD: Visible Only false | The Plane is present in the file but deactivated. The 5 Cubes are present and active/visible. |

Particle system: Plane emitting 4 Cubes
| Setup | Result |
| -- | -- |
| Plane: Show Emitter true + USD: Visible Only true | The Plane and the 5 Cubes are present in the file and active/visible |
| Plane: Show Emitter true + USD: Visible Only false | The Plane and the 5 Cubes are present in the file and active/visible |
| Plane: Show Emitter false + USD: Visible Only true | Only the 5 Cubes are present in the file and active/visible |
| Plane: Show Emitter false + USD: Visible Only false | The Plane is present in the file but deactivated. The 5 Cubes are present and active/visible. |