Fix #144542: Workbench DoF crash when using Vulkan
The depth of field shader create info contained a resource that wasn't
used in the shader or bound. In the Vulkan backend the shader resource
list is used in the outer loop to find out which resources needs to be
requested. As this resource didn't exist it lead to issues down the
road.