iOS: Enable HDR/EDR rendering support
This patch enables support for HDR display on iOS by activating it
on the MetalKit View CAMetalLayer following the existing macOS
implementation (see #105662). The WWDC 2022 talk "Display HDR video in
EDR with AVFoundation and Metal"[^1] was also used as an additional
reference.

As the CAMetalLayer `wantsExtendedDynamicRangeContent` property is only
available on iOS 16+[^2], the iOS deployment target has been bumped from
15.0 to 16.0. Additionally, CoreGraphics had to be added to the list of
linked framework.

[^1]: https://developer.apple.com/videos/play/wwdc2022/110565
[^2]: https://developer.apple.com/documentation/quartzcore/cametallayer/wantsextendeddynamicrangecontent?language=objc