Color Management: Improve wide gamut color buttons and pickers
These should at least show the correct colors and gamut now, following the
color picking space in the OpenColorIO config.

However the ACES config use sRGB as the color_picking role, which means
that the color picker only shows a limited a gamut. There is apparently is
not agreement on the interpretation of this role, so we might have to stop
using it.

Additionally, if you set the color_picking role to something like ACEScg +
sRGB transfer function, the widgets do show all the colors in the gamut,
however the gamut is probably too large to be practical for most uses. So
we need some UI solution for this regardless.

Ref #144911

Pull Request: https://projects.blender.org/blender/blender/pulls/144565