Color Management: Add Rec.2100-PQ display and ACES 2.0 view transforms
These are built into OpenColorIO, and will start working well with the upcoming
wide gamut display changes.

The ACES 2.0 view transform provides an alternative to AgX and Filmic, with
similar purpose. It is also useful for those working in an ACES pipeline when
the full ACES OpenColorIO config is not needed.

Rec.2100-PQ display is a HDR display, and will be what you select when wanting
to see HDR (and wide gamut) colors, and export them to HDR video.

Pull Request: https://projects.blender.org/blender/blender/pulls/144565