Direcly compute the masking for corner pin
As it has the derivatives of the sample location, it is pretty straight-forward
to compute the coverage of the source rectangle. There were artifacts at the
horizon line which took some work to eradicate.

This was copied to the anisotropic shader and the glsl shader to generate the
mask was removed.

The planetrackdeform node could also be fixed to directly compute the mask,
however I did not change it as I have no good sample data to test it with.