Fix #144383: Limited dissolve creates duplicate faces
Use of BM_faces_join_pair can result in an invalid mesh with doubled
faces. [0] added an assert to identify when this could occur.

This case has been inspected, and allowing the auto-join logic in
BM_faces_join_pair() to delete the doubled face does not negatively
impact iteration or processing:

1. The eheap and eheap_table used here deal only with edges, not faces.
2. Loop iteration is unchanged.
   A conversion to use of BM_ITER_MESH_MUTABLE is not applicable because
   the iteration is done on the edge heap, not the mesh.
3. The recomputation of the face normals of each combined face still
   works properly.

Ref !144653

[0]: 702efd6846790cd820e7ef20b81500eb21d902cd