UI: Use UI_alert For Vulcan Fallback Warning
This PR just updates the error message shown when the selected backend
is Vulcan but that fails and uses OpenGL instead. This currently uses a
bespoke popup alert block, but this PR replaces that with new UI_alert
function with the same wording.

Pull Request: https://projects.blender.org/blender/blender/pulls/144676