Compositor: Add Convert to Display node
This applies an OpenColorIO display, view and look transform on a color
in the scene_linear colorspace.

In general, OpenColorIO configurations do not contain a colorspace for
every view + display, especially if they are modern configs using the
display colorspace and shared view mechanisms. Nor do they include looks.
So the Convert Colorspace node is not sufficient.

Additionally, we would like to avoid making the colorspace list too long
in the default config, as we are adding many new views and transforms.

---

One design issue is that there are hidden gamma, exposure, white point
settings that don't make much sense to expose in the compositor, because
there are dedicated nodes for that. However these are technically editable
through the Python API.

Putting just the properties we need in a new data structure will give a
lot of code duplication, since there is reasonably complicated logic about
how to keep these properties in sync, do versioning, etc. Is this acceptable
as is?