Nodes: Don't delete nodes when cancelling swap operator
Use `node.translate_attach` instead of `node.translate_attach_remove_on_cancel` for the Swap Node and Zone operators.

The latter makes sense with Add node, as removing the node puts the nodetree back in a state identical to before the operator is called.

This doesn't apply when swapping nodes, and while it's not currently possible to bring the nodetree to the state before the swap when cancelling the operator, keeping the newly swapped node in the same position is a decent UX compromise.