iOS: Edge Swipe
Supported gestures:
* Inwards left edge.
* Inwards right edge.

This calls an operator (screen.edge_swipe) which is supposed to handle
all edge swipes. At the moment it only works for the viewport to bring
the Tools and Properties Region in when in Focus Mode.

I could not get outwards edge swipe to work. It requires a regular pan
gesture, and it fights with other gestures (including the edge swipe).