Compositor: Add Jitter input to Glare Sun beams
This patch introduces a new Jitter option for the Sun Beams mode of the
Glare node. If non-zero, is uses an approximation that is faster to
compute but makes the result more noisy.

Internally, this is computed by reducing the number of samples used to
integrate the results and introducing a random offset during
integration.

Pull Request: https://projects.blender.org/blender/blender/pulls/141789