Color Management: Base color space interpretation on interop ID
For display and image saving, we need to know certain properties of displays
and view transforms that OpenColorIO does not explicitly provide. In the
upcoming OpenColorIO 2.5 there will be an interop ID for color spaces that
adds more information. This moves towards that design, with heuristics for
configs that don't have the interop IDs.

* Try to use the first colorspace alias as the interop ID. This is used in
  some existing configs, and now the Blender config as well.
* Improve ACES 2.0 config compatibility by interpreting some aliases like
  srgb_display as interop IDs.
* Detect is_srgb and is_wide_gamut through display colorspace interop ID.
  These are now per view rather than per display.
* Detect is_extended where we can't clamp to 0..1 through interop ID.

* Detect untonemapped view as config wide default view transform for converting
  between reference and display space, if "Standard" or "Un-tone-mapped" can
  not be found.
* Detect which display and view transform combination is HDR by checking for
  "hdr-video" encoding on the display colorspace in the OpenColorIO config.

Pull Request: https://projects.blender.org/blender/blender/pulls/144565