Geometry Nodes: implement file path callback for import nodes
Add geometry file import nodes support to the for-each-path logic. This
will make `bpy.data.file_path_map()` report the input files for OBJ,
PLY, etc. import nodes.

Currently this only supports static file paths, so where the file path
is set as the input socket's default value. When the path is determined
via any noodle, this is ignored and the default value is reported
anyway.

This is necessary for the new version of Blender Asset Tracer, which in
turn is needed to resolve studio/flamenco#104423.