EEVEE: World Sun Shadow no longer works in 4.5
On some drivers, the GLSL compiler doesn't reflect the omitted
`local_size_*` of a compute shader inside `gl_WorkGroupSize`.

This lead to the 2D size computation of 1D workgroups to become
0 which was bypassing the parallel reduction algorithms.

Ensuring `local_size_*` are always set fixes the issue.

For clarity, also fix the 1D shaders to not use `gl_WorkGroupSize.y`.
This also fix a copy paste error in the Metal backend.

This issue affected AMD drivers on Windows.

Rel #142046

Backport of #144056

# Conflicts:
#	source/blender/draw/engines/eevee_next/shaders/eevee_lightprobe_sphere_irradiance_comp.glsl
#	source/blender/draw/engines/eevee_next/shaders/eevee_lightprobe_sphere_sunlight_comp.glsl
#	source/blender/gpu/intern/gpu_shader_create_info.hh