Stabilization for stretch constraint rotations.
Stretch/Shear constraint rotation offsets rely on the segment length remaining
close to the rest length for correct angles. This is not always the case, for
example when using pinned points and fast movement with stiff constraints.
The stretch constraint angle changes become unstable when a segment is stretched
to a multiple of its rest length.

To fix this the residual is scaled to represent the correct chord on the unit
sphere. That way the lambda vector can be interpreted as a rotational axis again
with the correct angle for aligning the rotation z axis to the segment.