Vulkan: Reduce memory pressure when uploading data
This PR reduces the memory pressure when uploading data to the GPU. In
the previous situation uploading data would be part of the context
rendergraph and would start uploading the data when the context is
flushed. This is now changed by creating a render graph with only the
upload part and submit it immediately.

The pressure is reduced as the staging buffers can be freed when the
upload has been completed and doesn't need to wait for all the other
commands to be completed.