Vulkan: User pref to disable fallback to OpenGL
When Blender starts it creates a temporarily Vulkan instance to detect
if the platform (OS, GPU, Driver) is supported and all required extensions
are present. When not Blender will fallback to OpenGL.

There are issues on certain platforms that the Driver doesn't perform a
clean exit and doesn't allow a second Vulkan instance to be created.