Fix for unstable bending constraints.
The formulation of offsets was based on the older 2016 paper
"Position and Orientation Based Cosserat Rods" where the Darboux vector
and lambda are 3D vectors and are inverted for one of the point offsets.
With the inclusion of the real component of the Darboux vector in
"Sag-Free Initialization for Strand-Based Hybrid Hair Simulation" (Hsu 2022)
the second offset instead is the conjugate of a quaternion, only the imaginary
part is inverted.

This was causing unbalanced rotations at higher bending angles, which was
introducing energy into the system in the form of accelerating rotations.