Nodes: Add "Swap" menu to Node Editor header
This makes it so that the F3 operator search displays the menu entries as "Swap (Shift + S) > ..." (similar to "Add (Shift + A)" for the Add Menu). As normally, menu entries do not show hotkeys in the search.

This also has the added benefit of the Swap menu no longer being weighed higher than Add in search. This is due to the hotkey text for the Add menu making the search entries longer, thus putting them lower in the search list.