Color Management: Improve gamma 2.2 vs piecewise sRGB handling for HDR
See the code comments for detailed explanation. But overall we switch sRGB
to gamma 2.2 encode in some cases to ensure it cancels out exactly with
the gamma 2.2 decode in the Vulkan HDR swapchain and values are passed on
linearly for accurate HDR display.

This means the user interface and SDR content will display different in
Windows HDR mode off and on, but that is consistent with other applications.

Split use_hdr variable into use_hdr_buffer about the graphics buffer,
and use_hdr_display about the chosen display in Blender.

Ref #145022

Pull Request: https://projects.blender.org/blender/blender/pulls/145025