Multires: Add 'Conform Base' functionality
This commit modifies the `OBJECT_OT_multires_base_apply` operator to
take in a new parameter which determines whether the final heuristic of
adjusting for a subdivsiion surface modifier applies.

Resolves #124553

---

Note, currently the tooltip in the modifier menu is the same for both
entries. If we think this is cruicial before landing this, the effort
to change the tooltip will likely need some minor general UI work.