Fix: Nodes: Swap operator casting to bool does not match implicit conversion behavior
Implicitly converting an Integer, Float, or Color to a Boolean is equivalent to a greater-than 0 operation.

However, Python's bool` constructor returns True any time the value isn't equivalent to 0, so `cast_value` is updated to account for this.

An implicit conversion from Vector to Boolean does match the default `bool` constructor behavior, so that is kept specifically for that case.