Fix #144625: Vulkan: Blacklist NVIDIA 580.76.05 on Linux
This PR disables support for NVIDIA 580.76.05 drivers as on many
platforms it hangs on startup (Wayland). There are reports from other
applications that this driver is faulty and NVIDIA confirmed that an
updated driver is in the making.

For now we will blacklist this specific driver on any Linux system.