Fix #144864: UI freezes with lots of Action slots (in the Outliner) part 2
Fix an omission in my earlier fix (50876d9b26839). This resolves the
situation for the "Blend File" mode, where Actions are shown both in a
flat list (which should show all slots) as well as underneath the IDs
they animate (which should only show the assigned slot).

Instead of looking at the Outliner mode, the code now uses the type of
the parent tree element. If this is an "Animation" element, a single
slot is shown. Otherwise all slots are shown.