Reduce cost of DupliObject `random_id` generation
Each dupli is given a PersistentId which is 8 integers, 1 for each level
of recursive duplication that Blender supports. When an instance belongs
to fewer than 8 levels, the id is padded with sentinal values for the
remaining levels. For example, a 2 level dupli may have a persistent id
like:`[7, 4, INT_MAX, INT_MAX, INT_MAX, INT_MAX, INT_MAX, INT_MAX]`

When calculating the instance's `random_id` we were previously
considering all 8 values of the Persistent Id which is unnecessary and
wasteful. This change does have the downside that the Object Info:Random
material socket changes values from prior versions.

This improves the performance of typical scenes by roughly 4-6%. The
Flower scatter demo w/PR 140378 applied lands in the middle of that
range.