Moved RealizationOptions to blender::math::SamplingOptions
Renamed:
   Interpolation to Sampler
   Bilinear to Box
   Bicubic to Bspline
   ExtensionMode to InterpWrapMode
(user interface and DNA not changed)

This structure is passed to all the sampling functions in the compositor,
and most nodes have a get_options() to return this rather than 3 functions
to return the interpolation and wrap modes.

This is in preparation for making a C++ sample_rect() function in blender::math.