Fix #143802: Grease Pencil: Divide blend mode not working in viewport
For the "Divide" layer blend mode to work in the viewport, we
need to use a floating point texture format for rendering.

This changes the format to always use floats for the `color` textures,
but still switches the `reveal_tx` to use integers in the viewport, and
only floating point during final render.

Pull Request: #143993