Fix pose assets activated when spawning context menu
Add explicit flag to enable that behavior. It's needed for brushes
still. Also still highlight the asset as active when spawning the
context menu, so it's clear which asset the menu belongs too.

Also fix previous active item being drawn as selected when spawning a
context menu from a different one.