Fix #145287: Don't touch original mesh when converting to grease pencil
Previously `BKE_id_material_clear` is used on the original mesh data for
clearing mesh array when converting mesh objects into grease pencil,
this is not desired because:

1. It damages original mesh data block, if the mesh is shared across
   multiple objects, this can lead to unwanted changes.
2. Converting multiple mesh objects that shares one mesh can crash
   because the now-modified mesh data block is not fully evaluated and
   subsequent conversions aren't able to get a valid mesh.

By removing the `BKE_id_material_clear` call we can make the conversion
work as expected. Since we replace the object data and reset material
count in the object, the material array in the original mesh shouldn't
pose any influence on the result of the conversion.