vulkan: Let WSI implementation define swapchain count
If we choose minImageCount, we will generally only have 1 free frame to render to
this means even if we are using mailbox mode, we can block occasionally waiting for a frame to be free.

This solves frame pacing issues on Wayland, and potentially other WSI
implementations, even when GHOST_FRAMES_FLIGHT is set to 5.