Fix #145314: Vulkan: Disable HDR on incapable drivers
This PR disables HDR when driver doesn't support
`VK_EXT_swapchain_colorspace`. It could be that Windows detect support
(DirectX) and allow users to enable HDR. But that it isn't supported by
the Vulkan part of the driver.

In case of Windows on Arm there might be an issue with using SRGB non
linear on F16 swapchains where somewhere an additional transform is
being performed. When we add support for other color spaces we should
never select this configuration.