UI: Option to Always Show Number Arrows
Numerical inputs have little arrows on the left and right to allow
quick increment/decrement of the value. But these are only shown on
hover, which does help if using a device that does not have a hover
state (some pens, most touch). This PR adds a user preference, off by
default, to show these always.