FIx: Nodes: `use_transform` not properly working if node is excluded via poll
If a node operator has a poll function and it is false, it is not drawn in the node menus and returns `None` for operator properies. So errors are created when trying to apply `props = use_transform` if `props` is `None`.

To fix this, a `set_use_transform` helper is created to check and only apply `use_transform` if it's valid to do so.