Fix: Nodes: Error when casting transferring values of menu sockets with Swap Node operator
This happens because the sockets are typically named `type`. So a value transfer is attempted more often even if the enums are not matching (i.e Filter & Glare types). It's hard to make an explicit check for this, so we just catch on a `TypeError` instead as its likely triggered by this.