Color Management: Change Blender config to use display color spaces
Use the OpenColorIO v2 to transform to cie_xyz_d65_interchange, and
then to the display colorspace.

The inverse AgX LUT was remove, since it got a bit inconvenient to
keep in this new framework, and it never really worked that well
anyways.

Pull Request: https://projects.blender.org/blender/blender/pulls/146487