Fix #144636: EEVEE could raytrace invalid samples
Raytracing pipeline in EEVEE interpolates in time for better results.
But depending on the interpolation could try to load a sample that isn't
available. This PR fixies this for planar probes. Screen tracing doesn't
seem to be effective (or the artifacts are not visible as the input
radiance is always known).